+++ packages/toolpad-app/src/components/AppEditor/PagePanel.tsx
@@ -1,4 +1,4 @@
-import { styled } from '@mui/material';
+import { styled, SxProps } from '@mui/material';
 import * as React from 'react';
 import HierarchyExplorer from './HierarchyExplorer';
 
@@ -9,11 +9,12 @@ const PagePanelRoot = styled('div')({
 export interface ComponentPanelProps {
   appId: string;
   className?: string;
+  sx?: SxProps;
 }
 
-export default function PagePanel({ appId, className }: ComponentPanelProps) {
+export default function PagePanel({ appId, className, sx }: ComponentPanelProps) {
   return (
-    <PagePanelRoot className={className}>
+    <PagePanelRoot className={className} sx={sx}>
       <HierarchyExplorer appId={appId} />
     </PagePanelRoot>
   );
